22 system autoryzacji i autentykacji

+++ backend/ezgloszenie/src/main/java/io/github/pluton33/ezgloszenie/service/ReportServiceImpl.java
@@ -1,11 +1,8 @@
 package io.github.pluton33.ezgloszenie.service;
 
-import io.github.pluton33.ezgloszenie.data.Report;
-import io.github.pluton33.ezgloszenie.data.ReportEntity;
-import io.github.pluton33.ezgloszenie.data.ReportsResponse;
+import io.github.pluton33.ezgloszenie.data.*;
 import io.github.pluton33.ezgloszenie.repository.ReportsRepository;
-import org.modelmapper.ModelMapper;
-import org.springframework.beans.factory.annotation.Autowired;
+import io.github.pluton33.ezgloszenie.repository.UsersRepository;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.web.server.ResponseStatusException;
@@ -14,71 +11,81 @@ import java.util.List;
 
 @Service
 public class ReportServiceImpl implements ReportService {
-    @Autowired
-    private ReportsRepository reportsRepository;
+    private final ReportsRepository reportsRepository;
+    private final UsersRepository usersRepository;
+    private final ReportMapper reportMapper;
 
-    @Autowired
-    private ModelMapper modelMapper;
+    public ReportServiceImpl(ReportsRepository reportsRepository, UsersRepository usersRepository, ReportMapper reportMapper) {
+        this.reportsRepository = reportsRepository;
+        this.usersRepository = usersRepository;
+        this.reportMapper = reportMapper;
+    }
 
     @Override
     public ReportsResponse getReports() {
         List<ReportEntity> reportEntities = reportsRepository.findAll();
         List<Report> reports = reportEntities.stream()
-                .map(entity -> Report.fromEntity(entity))
+                .map(entity -> reportMapper.toDto(entity))
                 .toList();
         return new ReportsResponse(reports);
     }
 
     @Override
-    public Report getReportById(int id) {
+    public Report getReportById(long id) {
         return reportsRepository
                 .findById(id)
-                .map(entity -> Report.fromEntity(entity))
+                .map(entity -> reportMapper.toDto(entity))
                 .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Brak zgłoszenia"
                 ));
     }
 
     @Override
-    public Report addReport(Report report) {
-        ReportEntity reportEntity = report.toEntity();
+    public Report addReport(Report report, String email) {
+        UserEntity loggedUser = usersRepository.findByEmail(email).orElseThrow(
+                () -> new ResponseStatusException(HttpStatus.NOT_FOUND, "User not found"));
+        ReportEntity reportEntity = reportMapper.toEntity(report);
         reportEntity.setId(null);
+        reportEntity.setUser(loggedUser);
         ReportEntity savedEntity = reportsRepository.save(reportEntity);
 
-        return new Report(
-                savedEntity.getId(),
-                savedEntity.getTitle(),
-                savedEntity.getContent()
-        );
+        return reportMapper.toDto(savedEntity);
     }
 
     @Override
-    public Report editReport(int id, Report report) {
-        if(report.id() == null) {
+    public Report editReport(long id, Report report, String email) {
+        if (report.id() == null) {
             throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "ID is required");
         }
 
-        if(id != report.id()) {
+        if (id != report.id()) {
             throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "ID mismatch");
         }
+        //Fetchuje po id użytkownika i sprawdza czy podany email zgadza się z tym w reporcie sprzed edycji
+        ReportEntity oldReportEntity = reportsRepository.findById(id).
+                orElseThrow(()->new ResponseStatusException(HttpStatus.NOT_FOUND, "Report not found"));
 
-        if(!reportsRepository.existsById(report.id())){
-            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Report not found");
+        if (!oldReportEntity.getUser().getEmail().equals(email)) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "No editing privileges for this report");
         }
-        ReportEntity reportEntity = report.toEntity();
+        ReportEntity reportEntity = reportMapper.toEntity(report); //do korekty żeby nie tworzyć nowego obiektu reportEntity
         reportEntity.setId(id);
+        reportEntity.setUser(oldReportEntity.getUser());
         ReportEntity editedEntity = reportsRepository.save(reportEntity);
 
-        return new Report(
-                editedEntity.getId(),
-                editedEntity.getTitle(),
-                editedEntity.getContent()
-        );
+        return reportMapper.toDto(editedEntity);
     }
 
     @Override
-    public void deleteReport(int id) {
-        if(!reportsRepository.existsById(id)) {
+    public void deleteReport(long id, String email) {
+        if (!reportsRepository.existsById(id)) {
             throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Report not found");
+        }
+
+        ReportEntity oldReportEntity = reportsRepository.findById(id).
+                orElseThrow(()->new ResponseStatusException(HttpStatus.NOT_FOUND, "Report not found"));
+
+        if (!oldReportEntity.getUser().getEmail().equals(email)) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "No editing privileges for this report");
         }
         reportsRepository.deleteById(id);
     }
